flask/requirements.txt
discord.py==2.4.0
Flask==2.3.3
Flask_Migrate==4.0.5
Flask_SocketIO==5.3.6
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
python-dotenv==1.0.1
Zenora==0.0.3.post1